+++ __main__.py
@@ -1,4 +1,4 @@
-from controllers.cli import EmsShell
+from interfaces.cli import EmsShell
 
 if __name__ == '__main__':
     EmsShell().cmdloop()
